biomni/requirements.txt
# Core dependencies
fastapi==0.115.6
uvicorn[standard]==0.24.0
pydantic>=2.5.2,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# JWT verification (Supabase Auth Bearer token)
PyJWT>=2.8.0

# AWS Bedrock
boto3==1.41.5

# Lambda (Mangum adapter for FastAPI on API Gateway)
mangum>=0.17.0

# Utilities
python-dotenv==1.0.0
urllib3>=2.0.0,<3.0.0
httpx>=0.24.0,<0.25.0
websockets>=11.0,<13.0

# Logging
structlog==23.2.0

# Biomni biomedical agent
biomni>=0.0.8
langchain-aws==0.2.35
langchain-openai>=0.2.0
reportlab>=4.0.0
pandas>=2.0.0
